requirements.txt
Flask
pdf2image
pandas
opencv-python
Pillow
google-cloud-vision
numpy
xlsxwriter
